+++ tests/test_column_affinity.py
@@ -1,5 +1,5 @@
 import pytest
-from sqlite_minutils.utils import column_affinity
+from apswutils.utils import column_affinity
 
 EXAMPLES = [
     # Examples from https://www.sqlite.org/datatype3.html#affinity_name_examples